refactor: move Commit to its own file
